moved and unified primitives code
